Move FileMenu *.cpp *.h files to belonging folder
- rename backstage folder into filemenu
- strip backstage from filenames
- make recent files working again
