+++ JASP-Desktop/widgets/filemenu/firsttabsstylesheet.qss
