+++ JASP-Desktop/widgets/filemenu/secondtabsstylesheet.qss
